+++ lib/views/dashboard/widgets/widgets.dart
@@ -1,3 +1,4 @@
+export 'dashboard_title.dart';
 export 'intranet_ip.dart';
 export 'network_detection.dart';
 export 'network_speed.dart';